+++ crdesigner/map_conversion/opendrive/opendrive_conversion/plane_elements/traffic_signals.py
@@ -5,7 +5,7 @@ from crdesigner.map_conversion.opendrive.opendrive_conversion import utils
 from typing import Union, Tuple, List
 
 from crdesigner.map_conversion.opendrive.opendrive_parser.elements.road import Road
-from crdesigner.map_conversion.common.utils import generate_unique_id
+from crdesigner.map_conversion.common.utils import generate_unique_id, get_default_cycle
 from crdesigner.map_conversion.opendrive.opendrive_parser.elements.roadLanes import LaneSection
 from crdesigner.map_conversion.opendrive.opendrive_parser.elements.roadSignal import Signal
 
@@ -162,7 +162,8 @@ def get_traffic_signals(road: Road) -> Tuple[List[TrafficLight], List[TrafficSig
             # we ignore such signals in order not cause trouble in traffic simulation
             if signal.type != ("1000002" or "1000007" or "1000013"):
 
-                traffic_light = TrafficLight(traffic_light_id=signal.id + 2000, position=position)
+                traffic_light = TrafficLight(traffic_light_id=signal.id + 2000, position=position,
+                                             traffic_light_cycle=get_default_cycle())  # TODO remove for new CR-Format
 
                 traffic_lights.append(traffic_light)
             else: